requirements.txt
wheel==0.36.2
pip==21.0.1
keras==2.4.3
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0
streamlit==0.75.0
tensorflow==2.4.1